pyproject.toml
[project]
name = "expense-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.5.0",
    "notebook>=7.5.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "seaborn>=0.13.2",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
